+++ compiler-core/src/parse.rs
@@ -949,10 +949,11 @@ where
 
     // An assignment, with `Let` already consumed
     fn parse_assignment(&mut self, start: u32) -> Result<UntypedStatement, ParseError> {
-        let kind = if let Some((assert_start, Token::Assert, assert_end)) = self.tok0 {
+        let mut kind = if let Some((assert_start, Token::Assert, assert_end)) = self.tok0 {
             _ = self.next_tok();
             AssignmentKind::Assert {
                 location: SrcSpan::new(assert_start, assert_end),
+                message: None,
             }
         } else {
             AssignmentKind::Let
@@ -985,11 +986,22 @@ where
                 },
             },
         })?;
+
+        let mut end = value.location().end;
+
+        match &mut kind {
+            AssignmentKind::Let | AssignmentKind::Generated => {}
+            AssignmentKind::Assert { message, .. } => {
+                if self.maybe_one(&Token::As).is_some() {
+                    let message_expression = self.expect_expression_unit()?;
+                    end = message_expression.location().end;
+                    *message = Some(Box::new(message_expression));
+                }
+            }
+        }
+
         Ok(Statement::Assignment(Assignment {
-            location: SrcSpan {
-                start,
-                end: value.location().end,
-            },
+            location: SrcSpan { start, end },
             value: Box::new(value),
             pattern,
             annotation,
@@ -1556,13 +1568,20 @@ where
 
             Some((start, Token::Name { name }, end)) => {
                 self.advance();
-                let mut unit = ClauseGuard::Var {
-                    location: SrcSpan { start, end },
-                    type_: (),
-                    name,
-                };
 
                 self.parse_function_call_in_clause_guard(start)?;
+
+                let mut unit = if let Some(record) =
+                    self.parse_record_in_clause_guard(&name, SrcSpan { start, end })?
+                {
+                    record
+                } else {
+                    ClauseGuard::Var {
+                        location: SrcSpan { start, end },
+                        type_: (),
+                        name,
+                    }
+                };
 
                 loop {
                     let dot_s = match self.maybe_one(&Token::Dot) {
@@ -1640,6 +1659,37 @@ where
                     Ok(None)
                 }
             }
+        }
+    }
+
+    fn parse_record_in_clause_guard(
+        &mut self,
+        module: &EcoString,
+        module_location: SrcSpan,
+    ) -> Result<Option<UntypedClauseGuard>, ParseError> {
+        let (name, end) = match (self.tok0.take(), self.peek_tok1()) {
+            (Some((_, Token::Dot, _)), Some(Token::UpName { .. })) => {
+                self.advance(); // dot
+                let Some((_, Token::UpName { name }, end)) = self.next_tok() else {
+                    return Ok(None);
+                };
+                (name, end)
+            }
+            (tok0, _) => {
+                self.tok0 = tok0;
+                return Ok(None);
+            }
+        };
+
+        if let Some(record) = self.parse_const_record_finish(
+            module_location.start,
+            Some((module.clone(), module_location)),
+            name,
+            end,
+        )? {
+            Ok(Some(ClauseGuard::Constant(record)))
+        } else {
+            Ok(None)
         }
     }